requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
kafka-python==2.0.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
mlflow==2.8.1
prometheus-client==0.19.0
python-dotenv==1.0.0
requests==2.31.0
evidently==0.4.14

